+++ Braintree/src/sharedTest/java/com/braintreepayments/testutils/FixturesHelper.java
@@ -8,6 +8,7 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.UnsupportedEncodingException;
+import java.util.Base64;
 
 import static androidx.test.InstrumentationRegistry.getTargetContext;
 
@@ -22,6 +23,18 @@ public class FixturesHelper {
                 return stringFromAndroidFixture(filename);
             } catch (RuntimeException | FileNotFoundException | Error e) {
                 return stringFromUnitTestFixture(filename);
+            }
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public static String base64EncodedClientTokenFromFixture(String filename) {
+        try {
+            try {
+                return Base64.getEncoder().encodeToString(stringFromAndroidFixture(filename).getBytes());
+            } catch (RuntimeException | FileNotFoundException | Error e) {
+                return Base64.getEncoder().encodeToString(stringFromUnitTestFixture(filename).getBytes());
             }
         } catch (IOException e) {
             throw new RuntimeException(e);