+++ Demo/src/main/java/com/braintreepayments/demo/models/PayPalUAT.java
@@ -1,0 +1,10 @@
+package com.braintreepayments.demo.models;
+
+import com.google.gson.annotations.SerializedName;
+
+public class PayPalUAT {
+    @SerializedName("universal_access_token")
+    private String mUAT;
+
+    public String getUAT() { return mUAT; }
+}