+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## unreleased
 
 * Allow new BraintreeFragment instances to be created using FragmentActivity
+* Add support for authorizing the Braintree SDK with a `PayPalUAT` (universal access token)
+* Fix bug that accepted raw JSON string as valid authorization to `BraintreeFragment.newInstance(...)`
 
 ## 3.9.0